refactor: 가독성을 위해 switch사용
